insurgig-main/insurgig-main/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.28.1
python-multipart==0.0.17
numpy>=2.1.0
pandas>=2.2.0
scikit-learn>=1.5.0
xgboost>=2.1.0
lightgbm>=4.5.0
joblib==1.4.2
networkx==3.2.1
python-louvain==0.16
PyJWT==2.10.1
